page thymleaf avec hello world

+++ appec/src/main/java/fr/univlyon1/m1if/m1if10/appec/dao/Dao.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface Dao<T> {
 
-    Optional<T> get(long id);
+    Optional<T> get(Integer id);
 
     List<T> getAll();
